Implement metal item model and palette loader
Armor trim is nearly perfect for this, it just sucks for addons. Implement my own variant of it that loads variants from a tag-like file
Might want to toss in some mod ID filtering on that variant file eventually; ensure mod is loaded before using it. But probably not needed