Add more metal effects
Tin now grants reach, and bronze energy
Iron now increases knockback taken, and changes dall distance in accordance to gravity change
Pewter and steel both affect mining speed now. Not 100% settled on which will keep that in the end, perhaps both as its a mixture of strength and speed
Rose gold now does health boost. This is also one I'm not fully set on as temporary health boost is a little awkward, we will see.
As part of this, also lowered the spike charge needed to 10 by default