Add support for melting and casting metal items in Tinkers'
Adds 3 new recipe types when tinkers is loaded: metal basin casting, metal table casting, and metal melting.
The new recipe types are used to provide melting and casting for bracers, rings, and spikes, plus lerasium nugget composite.
As part of this, the lerasium alloy recipe only returns 1 nugget instead of 2 so casting tables can work. Fits better with the rework for spikes.

This features a new ingredient type of MetalItemIngredient, which just displays metal variants on an item for JEI
It also includes a local copyof CastItemObject to register our custom cast items