Cleanup: split onTick into onTap and onStore
We already branched based on level in nearly every effect, and called each separately, so might as well be 2 methods.
Lets the storing value be positive in the activemetalminds logic, making it a bit easier to follow
onChange still uses the negative for store convention, as does MetalmindStack for its level field. Those make sense to leave as is based on their usages